+++ src/nationalarchives/components/hero/hero.stories.js
@@ -54,7 +54,6 @@ export default {
     ]),
   ),
   parameters: {
-    chromatic: { delay: 1000 },
     viewport: {
       options: customViewports,
     },