+++ src/nationalarchives/components/index-grid/index-grid.stories.js
@@ -43,9 +43,6 @@ export default {
       },
     ]),
   ),
-  parameters: {
-    chromatic: { delay: 1000 },
-  },
   render: (params) => {
     return nunjucks.renderString(Template, { params });
   },