+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Added `aria-current="page"` to current page in pages-style sidebar
+
 ### Security
 
 ## [0.32.0](https://github.com/nationalarchives/tna-frontend/compare/v0.31.0...v0.32.0) - 2026-01-21