+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.33.0...HEAD)
+## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.33.1...HEAD)
 
 ### Added
 ### Changed
@@ -13,6 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 ### Security
+
+## [0.33.1](https://github.com/nationalarchives/tna-frontend/compare/v0.33.0...v0.33.1) - 2026-02-10
+
+### Added
+
+- Added `tna-dl--left-weighted` and `tna-dl--right-weighted` as options to balance and size description list columns
+
+### Fixed
+
+- Fixed layout of description lists with very long content and balanced columns as 1:2
 
 ## [0.33.0](https://github.com/nationalarchives/tna-frontend/compare/v0.32.0...v0.33.0) - 2026-02-10